finished actions for UserController

+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\User;
+use JWTAuth;
+use Tymon\JWTAuth\Exceptions\JWTException;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
@@ -38,47 +40,22 @@ class UserController extends Controller
     public function store(Request $request)
     {
         $user = new User;
-        $password = Hash::make($request->input('password'));
-        $user->login = $request->input('login');
         $user->email = $request->input('email');
         $user->confirmed = false;
-        $user->password = $password;
-        $token = md5($request->input('email').time());
-        $confirm = new AccountConfirmation;
-        $confirm->token = $token;
-        $confirm->email = $request->input('email');
-        $host = $_ENV['USER_SVC_UI_SVC_URL'];
-        $link = $host.'/#/confirm?token='.$token;
-        $message = [
-            'to'      => $request->input('email'),
-            'subject' => 'Registration to acn-bootcamp.com',
-            'body'    => 'Dear '.$request->input('login').
-                         ', your registration at acn.bootcamp.com was successful!
-                          to confirm your account click on the following link:
-                          '.$link,
-        ];
-        $client = SqsClient::factory([
-                'key'    => $_ENV['USER_SVC_AWS_ACCESS_KEY'],
-                'secret' => $_ENV['USER_SVC_AWS_SECRET_KEY'],
-                'region' => $_ENV['USER_SVC_AWS_REGION'],
-            ]);
+        $user->password = bcrypt($request->input('password'));
+        $credentials = $request->only('email', 'password');
+
         try {
-            $client->sendMessage([
-                'QueueUrl'    => $_ENV['USER_SVC_AWS_EMAIL_QUEUE_URL'],
-                'MessageBody' => json_encode($message),
-            ]);
-        } catch (Exception $e) {
-            return response(['error' => 'Confirmation message send failed!'], 503);
+            if (!$token = JWTAuth::fromUser($user)) {
+                return response()->json(['error' => 'Invalid credentials'], 401);
+            }
+        } catch (JWTException $e) {
+            return response()->json(['error' => 'Could not create token'], 500);
         }
+
         $user->save();
-        $confirm->save();
-        error_log(json_encode($message));
-        return $user;
 
-        
-        // $user = User::create($request->all());
-
-        // return response()->json($user);
+        return response()->json($user);
     }
 
     /**
@@ -87,14 +64,17 @@ class UserController extends Controller
      * @param int $id
      * @return Response
      */
-    public function update($id)
+    public function update($id, Request $request)
     {
-        $user  = user::find($id);
+        $user = User::find($id);
         $user->name = $request->input('name');
         $user->surname = $request->input('surname');
-        $user->save();
+        if ($user->save()) {
+            return response()->json(['msg' => 'Success'], 200);
+        } else {
+            return response()->json(['error' => 'Invalid data'], 400);
+        }
  
-        return response()->json($user);
     }
 
     /**
@@ -105,8 +85,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::find($id);
+        if(!$user = User::find($id)) {
+            return response()->json(['error' => 'No such user'], 400);
+        }
         $user->delete();
-        return response()->json('success');
+        return response()->json(['msg' => 'Success'], 200);
     }
 }
